requirements.txt
tdl-client-python==0.27.13
coverage==7.5.4
pytest==8.2.2
pytest-cov==5.0.0
